requirements.txt
Pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0
